migrating firebase services to own authservices
