+++ lib/main.dart
@@ -1,9 +1,8 @@
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:mywaylearn/constants/routes.dart';
-import 'package:mywaylearn/firebase_options.dart';
+import 'package:mywaylearn/services/auth/auth_service.dart';
 import 'package:mywaylearn/view/login_view.dart';
+import 'package:mywaylearn/view/notes_view.dart';
 import 'package:mywaylearn/view/register_view.dart';
 import 'package:mywaylearn/view/verify_email_view.dart';
 import 'dart:developer' as devtools show log;
@@ -31,17 +30,15 @@ class HomePage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return FutureBuilder(
-      future: Firebase.initializeApp(
-        options: DefaultFirebaseOptions.currentPlatform,
-      ),
+      future: AuthService.firebase().initialize(),
       builder: (context, snapshot) {
         switch (snapshot.connectionState) {
           case ConnectionState.done:
-            final user = FirebaseAuth.instance.currentUser;
+            final user = AuthService.firebase().currentUser;
             // print(user);
             devtools.log(user.toString());
             if (user != null) {
-              if (user.emailVerified) {
+              if (user.isEmailVerified) {
                 return const NotesView();
               } else {
                 return const VerifyEmailView();
@@ -57,75 +54,3 @@ class HomePage extends StatelessWidget {
   }
 }
 
-enum MenuAction { logout }
-
-class NotesView extends StatefulWidget {
-  const NotesView({ Key? key }) : super(key: key);
-
-  @override
-  State<NotesView> createState() => _NotesViewState();
-}
-
-class _NotesViewState extends State<NotesView> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Main UI'),
-        backgroundColor: Colors.purple.shade600,
-        actions: [
-          PopupMenuButton<MenuAction>(
-            onSelected: (value) async {
-              // print (value);
-              switch (value) {
-                case MenuAction.logout:
-                  final shouldLogout = await showLogOutDiaLog(context);
-                  if (shouldLogout) {
-                    await FirebaseAuth.instance.signOut();
-                    Navigator.of(context).pushNamedAndRemoveUntil(loginRoute, (_) => false);
-                  }
-                  devtools.log(shouldLogout.toString());
-                  break;
-              }
-            },
-            itemBuilder: (context) {
-              return const [
-                PopupMenuItem<MenuAction>(
-                  value: MenuAction.logout,
-                  child: Text('Log out'),
-                ),
-              ];
-            },
-          ),
-        ],
-      ),
-      body: const Text('Notes is available'),
-    );
-  }
-}
-
-Future<bool> showLogOutDiaLog(BuildContext context) {
-  return showDialog<bool>(
-    context: context,
-    builder: (context) {
-      return AlertDialog(
-        title: const Text('Sign out'),
-        content: const Text('Are you sure you want to logout?'),
-        actions: [
-          TextButton(
-            onPressed: () {
-              Navigator.of(context).pop(false);
-            },
-            child: const Text('Cancel'),
-          ),
-          TextButton(
-            onPressed: () {
-              Navigator.of(context).pop(true);
-            },
-            child: const Text('Log Out'),
-          ),
-        ]
-      );
-    },
-  ).then((value) => value ?? false);
-}
