+++ lib/services/auth/auth_service.dart
@@ -1,10 +1,13 @@
 import "package:mywaylearn/services/auth/auth_user.dart";
 import "package:mywaylearn/services/auth/auth_provider.dart";
+import 'package:mywaylearn/services/auth/firebase_auth_provider.dart';
 
 class AuthService implements AuthProvider {
   final AuthProvider provider;
 
   const AuthService(this.provider);
+
+  factory AuthService.firebase() => AuthService(FirebaseAuthProvider());
 
   @override
   Future<AuthUser> createUser({
@@ -33,4 +36,7 @@ class AuthService implements AuthProvider {
 
   @override
   Future<void> sendEmailVerification() => provider.sendEmailVerification();
+
+  @override
+  Future<void> initialize() => provider.initialize();
 }
