+++ lib/view/login_view.dart
@@ -1,8 +1,8 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'dart:developer' as devtools show log;
 
 import 'package:mywaylearn/constants/routes.dart';
+import 'package:mywaylearn/services/auth/auth_exceptions.dart';
+import 'package:mywaylearn/services/auth/auth_service.dart';
 
 import '../utilites/show_error_dialog.dart';
 class LoginView extends StatefulWidget {
@@ -61,12 +61,12 @@ class _LoginViewState extends State<LoginView> {
               final email = _email.text;
               final password = _password.text;
               try {
-                await FirebaseAuth.instance.signInWithEmailAndPassword(
+                await AuthService.firebase().logIn(
                   email: email,
                   password: password,
                 );
-                final user = FirebaseAuth.instance.currentUser;
-                if(user?.emailVerified ?? false){
+                final user = AuthService.firebase().currentUser;
+                if(user?.isEmailVerified ?? false){
                   // user's email is verified
                   Navigator.of(context).pushNamedAndRemoveUntil(
                     notesRoute,
@@ -79,17 +79,12 @@ class _LoginViewState extends State<LoginView> {
                     (route) => false,
                   );
                 }
-              } on FirebaseAuthException catch (e) {
-                // devtools.log(e.code.toString());
-                if (e.code == 'user-not-found') {
-                  await showErrorDialog(context,'User is not Exist. Please register first.');
-                } else if (e.code == 'wrong-password') {
-                  await showErrorDialog(context,'Wrong Password. Please try again.');
-                } else {
-                  await showErrorDialog(context,'Error: ${e.code}');
-                } 
-              } catch (e) {
-                  await showErrorDialog(context,e.toString());
+              } on UserNotFoundAuthException {
+                await showErrorDialog(context,'User is not Exist. Please register first.');
+              } on WrongPasswordAuthException {
+                await showErrorDialog(context,'Wrong Password. Please try again.');
+              } on GenericAuthException {
+                await showErrorDialog(context,'Authentication error. Please try again.');
               }
             },
             child: const Text('Login'),
