+++ lib/view/register_view.dart
@@ -1,8 +1,9 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 // import 'dart:developer' as devtools show log;
 
 import 'package:mywaylearn/constants/routes.dart';
+import 'package:mywaylearn/services/auth/auth_exceptions.dart';
+import 'package:mywaylearn/services/auth/auth_service.dart';
 
 import '../utilites/show_error_dialog.dart';
 
@@ -62,23 +63,20 @@ class _RegisterViewState extends State<RegisterView> {
               final email = _email.text;
               final password = _password.text;
               try {
-                await FirebaseAuth.instance.createUserWithEmailAndPassword(email: email, password: password);
-                final user = FirebaseAuth.instance.currentUser;
-                await user?.sendEmailVerification();
+                await AuthService.firebase().createUser(
+                  email: email,
+                  password: password,
+                );
+                await AuthService.firebase().sendEmailVerification();
                 Navigator.of(context).pushNamed(verifyEmailRoute);
-              } on FirebaseAuthException catch (e) {
-                // devtools.log(e.code.toString());
-                if (e.code == 'weak-password') {
-                  await showErrorDialog(context,'To Weak Password.');
-                } else if (e.code == 'email-already-in-use') {
-                  await showErrorDialog(context,'Email alredy in use. Please try with a different email');
-                } else if (e.code == 'invalid-email') {
-                  await showErrorDialog(context,'Invalid Email Address');
-                } else {
-                  await showErrorDialog(context,'Error: ${e.code}');
-                }
-              } catch (e) {
-                  await showErrorDialog(context,e.toString());
+              } on WeakPasswordAuthException {
+                await showErrorDialog(context,'To Weak Password.');
+              } on EmailAlreadyInUseAuthException {
+                await showErrorDialog(context,'Email alredy in use. Please try with a different email');
+              } on InvalidEmailAuthException {
+                await showErrorDialog(context,'Invalid Email Address');
+              } on GenericAuthException {
+                await showErrorDialog(context,'Error while registering user!');
               }
             },
             child: const Text('Register'),
