+++ lib/view/verify_email_view.dart
@@ -1,6 +1,6 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:mywaylearn/constants/routes.dart';
+import 'package:mywaylearn/services/auth/auth_service.dart';
 
 class VerifyEmailView extends StatefulWidget {
   const VerifyEmailView({Key? key}) : super(key: key);
@@ -22,14 +22,13 @@ class _VerifyEmailViewState extends State<VerifyEmailView> {
           const Text("If you haven't received a verification email yet, press below button"),
           TextButton(
             onPressed: () async {
-              final user = await FirebaseAuth.instance.currentUser;
-              user?.sendEmailVerification();
+              await AuthService.firebase().sendEmailVerification();
             },
             child: const Text('Send Email Verification'),
           ),
           TextButton(
             onPressed: () async {
-              await FirebaseAuth.instance.signOut();
+              await AuthService.firebase().logOut();
               Navigator.of(context).pushNamedAndRemoveUntil(registerRoute, (route) => false);
             },
             child: const Text('Restart'),
